JRobin 1.3.1 DEBUG UPDATE - testing purposes - Bug fixes for bugs #0000032 and #0000033 - Improved Value Grid calculation - Improved Time Grid calculation for monthly gridlines - Fixed possible Trime Grid shift between minor and major lines - Modified right aligned title padding (signature overlap)

+++ src/org/jrobin/graph/ValueAxisUnit.java
@@ -225,17 +225,17 @@ class ValueAxisUnit
 		}
 		
 		double value		= ovalue * gridFactor;
-		int valueInt		= new Double(value).intValue();
-		int roundStep		= new Double(gridStep).intValue();
+		long valueInt		= new Double(value).longValue();
+		long roundStep		= new Double(gridStep).longValue();
 		if ( roundStep == 0 ) roundStep = 1;
-		int num 			= valueInt / roundStep; 
+		long num 			= valueInt / roundStep; 
 		// int mod 			= valueInt % roundStep;
 		double gridValue	= (roundStep * (num + 1)) * 1.0d;
 		if ( gridValue - value < (gridStep) / 8 )
 			gridValue		+= roundStep;
 		
 		value				= ovalue * mGridFactor;
-		roundStep			= new Double(mGridStep).intValue();
+		roundStep			= new Double(mGridStep).longValue();
 		if ( roundStep == 0 ) roundStep = 1;
 		num					= valueInt / roundStep;
 		// mod					= valueInt % roundStep;
